+++ CHANGES.rst
@@ -8,6 +8,7 @@ unreleased
 2016/11/18 0.4.0
 ================
 
+ - fix SMS text
  - mail extensions
    DEPLOY:
      ALTER TABLE petitions ADD COLUMN state['half_time_mail_time'] TIMESTAMP;